refactor: delete ci.rs and explain.rs, inline into main.rs
ci.rs (70 lines) was a string template for GitHub Actions YAML —
inlined into the codegen --ci handler.

explain.rs (81 lines) was a trivial wrapper around check::check()
that formatted markdown — inlined into the check --explain handler.

Two fewer modules, -106 net lines.
