--- crates/qedgen/src/explain.rs
@@ -1,81 +1,0 @@
-use anyhow::Result;
-use std::path::Path;
-
-use crate::check;
-
-/// Generate a Markdown verification report from spec and proof status.
-pub fn explain(spec_path: &Path, proofs_dir: &Path) -> Result<String> {
-    let results = check::check(spec_path, proofs_dir)?;
-
-    let spec_name = spec_path
-        .file_stem()
-        .map(|s| s.to_string_lossy().to_string())
-        .unwrap_or_else(|| "Spec".to_string());
-
-    let proven = results
-        .iter()
-        .filter(|r| r.status == check::Status::Proven)
-        .count();
-    let sorry = results
-        .iter()
-        .filter(|r| r.status == check::Status::Sorry)
-        .count();
-    let missing = results
-        .iter()
-        .filter(|r| r.status == check::Status::Missing)
-        .count();
-    let total = results.len();
-
-    let mut md = String::new();
-
-    // Header
-    md.push_str(&format!("# {} Verification Report\n\n", spec_name));
-
-    // Summary
-    md.push_str(&format!(
-        "**{}/{} properties verified** ({} sorry, {} missing)\n\n",
-        proven, total, sorry, missing
-    ));
-
-    if proven == total {
-        md.push_str("> All properties verified (sorry-free).\n\n");
-    }
-
-    // Properties table
-    md.push_str("## Properties\n\n");
-
-    for r in &results {
-        let icon = match r.status {
-            check::Status::Proven => "✓",
-            check::Status::Sorry => "✗",
-            check::Status::Missing => "✗",
-        };
-
-        let status_str = match r.status {
-            check::Status::Proven => "PROVEN",
-            check::Status::Sorry => "SORRY",
-            check::Status::Missing => "MISSING",
-        };
-
-        md.push_str(&format!("### {} {} — {}\n\n", icon, r.name, status_str));
-
-        if let Some(ref intent) = r.intent {
-            md.push_str(&format!("**Intent:** {}\n\n", intent));
-        }
-
-        if r.status != check::Status::Proven {
-            if let Some(ref suggestion) = r.suggestion {
-                md.push_str(&format!("**Suggestion:** {}\n\n", suggestion));
-            }
-        }
-    }
-
-    // Spec source
-    md.push_str("---\n\n");
-    md.push_str(&format!(
-        "*Generated by `qedgen explain` from `{}`*\n",
-        spec_path.display()
-    ));
-
-    Ok(md)
-}
